requirements.txt
astroid==2.3.3
autopep8==1.5.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.2
Pillow==7.0.0
pycodestyle==2.5.0
pylint==2.4.4
six==1.14.0
typed-ast==1.4.1
wrapt==1.11.2
